src/docproc-search-indexer/pyproject.toml
[project]
name = "docproc-search-indexer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.14",
    "azure-cosmos>=4.9.0",
    "azure-data-tables>=12.7.0",
    "azure-identity>=1.23.0",
    "azure-search-documents>=11.5.3",
    "openai>=1.54.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "tenacity>=8.2.0",
]
